backend/operation-apple-deployment-20250731_014024/requirements.txt
flask==3.1.1
flask-cors==3.0.10
flask-jwt-extended==4.6.0
gunicorn==21.2.0

# Core ML Dependencies
tensorflow==2.15.0
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
scikit-learn==1.3.2
opencv-python-headless==4.8.1.78
pillow==10.1.0
numpy==1.24.3
pandas==2.1.4

# 🍎 Operation Apple: Advanced Computer Vision Dependencies
opencv-contrib-python==4.8.1.78
scikit-image==0.21.0
scipy==1.11.4

# AI/ML Libraries
faiss-cpu==1.7.4
transformers==4.35.2
google-cloud-vision==3.4.4
google-auth==2.23.4
supabase==2.0.2

# Image Processing
Pillow==10.1.0
opencv-python-headless==4.8.1.78

# Performance and monitoring
psutil==5.9.6
memory-profiler==0.61.0

# Additional utilities
python-dotenv==1.0.0
requests==2.31.0 